LayoutTests/accessibility/ios-simulator/live-regions/live-region-types.html

<!DOCTYPE HTML><!-- webkit-test-runner [ IsAriaLiveRegionManagementEnabled=true ] -->
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/accessibility-helper.js"></script>
<meta http-equiv="content-language" content="en">
</head>
<body>

<div id="polite-region" aria-live="polite" aria-atomic="true">
    Price: $0
</div>

<div id="assertive-region" aria-live="assertive" aria-atomic="true">
    Count: 0
</div>

<script>
var output = "This test verifies that live region notifications are properly formatted on iOS devices.\n\n";
var notificationCount = 0;
var testFinished = false;

function notificationCallback(object, notification, userInfo) {
    if (notification == "AXAnnouncementRequested") {
        output += `Announcement received, notification: ${notification}, attributed string: ${userInfo["AXAnnouncementRequested"]}\n`;
        notificationCount++;

        if (notificationCount == 1)
            document.getElementById('assertive-region').textContent = "Count: 5";
        else if (notificationCount == 2)
            testFinished = true;
    }
}

if (accessibilityController) {
    window.jsTestIsAsync = true;
    accessibilityController.addNotificationListener(notificationCallback);
    accessibilityController.accessibleElementById("polite-region");

    document.getElementById('polite-region').textContent = "Price: $10";
 
    setTimeout(async function() {
        await waitFor(() => testFinished);
        accessibilityController.removeNotificationListener();
        debug(output);
        finishJSTest();
    }, 0);
}
</script>
</body>
</html>
