LayoutTests/accessibility/ios-simulator/live-regions/live-region-types-expected.txt

This test verifies that live region notifications are properly formatted on iOS devices.

Announcement received, notification: AXAnnouncementRequested, attributed string: Price: $10{
    UIAccessibilitySpeechAttributeAnnouncementPriority = UIAccessibilityPriorityLow;
    UIAccessibilitySpeechAttributeIsLiveRegion = 1;
    UIAccessibilitySpeechAttributeLanguage = en;
}
Announcement received, notification: AXAnnouncementRequested, attributed string: Count: 5{
    UIAccessibilitySpeechAttributeAnnouncementPriority = UIAccessibilityPriorityDefault;
    UIAccessibilitySpeechAttributeIsLiveRegion = 1;
    UIAccessibilitySpeechAttributeLanguage = en;
}

PASS successfullyParsed is true

TEST COMPLETE
Price: $10
Count: 5
